backend-fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
motor==3.3.2
pydantic==2.5.2
pydantic-settings==2.1.0
google-auth==2.23.4
PyJWT==2.8.0
google-generativeai==0.3.1
python-dotenv==1.0.0
google-genai
python-dotenv